+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Upgrade the `org.realityforge.react4j` artifacts to version `0.158`.
 * Upgrade the `org.realityforge.arez` artifacts to version `0.180`.
 * Upgrade the `org.realityforge.arez.testng` artifact to version `0.12`.
+* Explicitly specify `requireId = Feature.DISABLE` parameter on `@ArezComponent` annotations as the default behaviour will change in the next version of arez.
 
 ### [v6.90](https://github.com/replicant4j/replicant/tree/v6.90) (2020-05-26) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.89...v6.90)